+++ CHANGELOG.md
@@ -39,6 +39,8 @@
 * Fix crash in `RLMSuperSet` when assigning to a `RLMArray` property on a standalone object.
 * Add an error message when the protocol for an `RLMArray` property is not a
   valid object type.
+* Add an error message when an `RLMObject` subclass is defined nested within
+  another Swift class.
 
 0.86.3 Release notes (2014-10-09)
 =============================================================